requirements.txt
Pillow==8.4.0
requests==2.26.0
telepot==12.7
